refactor: stripp processing unrelated to channel
